SMM negative score (fix), exp. loops, UnicodePlots
