+++ src/GroupAD.jl
@@ -15,8 +15,7 @@ using Distributions
 include("data.jl")
 include("toy.jl")
 include("experiments.jl")
-include("experiments_point_cloud.jl")
+include("experimental_loops.jl")
 include("models/Models.jl")
-include("evaluation/Evaluation.jl")
 
 end #module
