+++ src/experimental_loops.jl
@@ -1,6 +1,141 @@
 """
+	edit_params(data, parameters)
+This modifies parameters according to data. Default version only returns the input arg. 
+Overload for models where this is needed.
+"""
+function edit_params(data, parameters)
+	parameters
+end
+
+"""
+	check_params(savepath, parameters)
+
+Returns `true` if the model with given parameters wasn't already trained and saved. 
+"""
+function check_params(savepath, parameters)
+	if ~isdir(savepath)
+		return true
+	end
+	# filter out duplicates created by tagsave
+	fs = filter(x->!(occursin("_#", x)), readdir(savepath))
+	# filter out model files
+	fs = filter(x->!(startswith(x, "model")), fs)
+	# if the first argument name contains a "_", than the savename is parsed wrongly
+	saved_params = map(x -> DrWatson.parse_savename("_"*x)[2], fs)
+	# now filter out saved models where parameter names are different or missing
+	pkeys = collect(keys(parameters))
+	filter!(ps->intersect(pkeys, Symbol.(collect(keys(ps))))==pkeys, saved_params)
+	for params in saved_params
+		all(map(k->params[String(k)] == parameters[k], pkeys)) ? (return false) : nothing
+	end
+	true
+end
+
+###################################################
+### MIL, point cloud and toy experimental loops ###
+###################################################
+
+"""
 	basic_experimental_loop(sample_params_f, fit_f, edit_params_f, 
-		max_seed, modelname, dataset, contamination, savepath, classes)
+		max_seed, modelname, dataset, contamination, savepath)
+
+This function takes a function that samples parameters, a fit function and a function that edits the sampled
+parameters and other parameters. Then it loads data, samples hyperparameters, calls the fit function
+that is supposed to construct and fit a model and finally evaluates the returned score functions on 
+the loaded data.
+
+Runs on MIL data.
+"""
+function basic_experimental_loop(sample_params_f, fit_f, edit_params_f, 
+		max_seed, modelname, dataset, contamination, savepath)
+	# set a maximum for parameter sampling retries
+	# this is here because you might sample the same parameters of an already trained model
+	# in that case this loop runs again, for a total of 10 tries
+	try_counter = 0
+	max_tries = 10*max_seed
+	while try_counter < max_tries
+		# sample the random hyperparameters
+	    parameters = sample_params_f()
+
+	    # with these hyperparameters, train and evaluate the model on different train/val/tst splits
+	    for seed in 1:max_seed
+	    	# define where data is going to be saved
+			_savepath = joinpath(savepath, "$(modelname)/MIL/$(dataset)/seed=$(seed)")
+			mkpath(_savepath)
+
+			# get data
+			@time data = load_data(dataset, seed=seed, contamination=contamination)
+			@info "Data loaded..."
+
+			# edit parameters
+			# bagkNN and SMM need to calculate parameter from the data
+			# the parameter needs to stay the same for all seeds
+			if modelname in ["bag_knn","SMM" ]
+				if seed == 1
+					parameters = edit_params_f(data, parameters)
+					edited_parameters = parameters
+				else
+					edited_parameters = parameters
+				end
+			else
+				# other models can edit parameters for each seed, because the change is deterministic
+				edited_parameters = edit_params_f(data, parameters)
+			end
+
+
+			@info "Trying to fit $modelname on $dataset with parameters $(edited_parameters)..."
+			@info "Train/validation/test splits: $(size(data[1][1], 2)) | $(size(data[2][1], 2)) | $(size(data[3][1], 2))"
+			@info "Number of features: $(size(data[1][1], 1))"
+
+			# check if a combination of parameters and seed alread exists
+			if check_params(_savepath, edited_parameters)
+				# fit
+				training_info, results = fit_f(data, edited_parameters)
+
+				# save the model separately			
+				if training_info.model != nothing
+					modelf = joinpath(_savepath, savename("model", edited_parameters, "bson", digits=5))
+					tagsave(
+						modelf, 
+						Dict("model"=>training_info.model,
+							"fit_t"=>training_info.fit_t,
+							"history"=>training_info.history,
+							"parameters"=>edited_parameters
+							), 
+						safe = true)
+					(@info "Model saved to $modelf")
+
+					training_info = merge(training_info, (model = nothing,history=nothing))
+				end
+
+				# here define what additional info should be saved together with parameters, scores, labels and predict times
+				save_entries = merge(training_info, (modelname = modelname, seed = seed, dataset = dataset))
+
+				# now loop over all anomaly score funs
+				@time for result in results
+					if modelname in ["vae_instance", "statistician", "PoolModel"]
+						experiment_bag(result..., data, _savepath; save_entries...)
+					elseif modelname == "SMM"
+						experiment(result..., GroupAD.Models.unpack_mill.(data), _savepath; save_entries...)
+					elseif modelname == "bag_knn"
+						experiment_bagknn(result..., GroupAD.Models.unpack_mill.(data), _savepath; save_entries...)
+					else # vae_basic, knn_basic
+						experiment(result..., data, _savepath; save_entries...)
+					end
+				end
+				try_counter = max_tries + 1
+			else
+				@info "Model already present, trying new hyperparameters..."
+				try_counter += 1
+			end
+		end
+	end
+	(try_counter == max_tries) ? (@info "Reached $(max_tries) tries, giving up.") : nothing
+end
+
+"""
+	point_cloud_experimental_loop(sample_params_f, fit_f, edit_params_f, 
+		max_seed, modelname, dataset, contamination, savepath, anomaly_classes, method)
 
 This function takes a function that samples parameters, a fit function and a function that edits the sampled
 parameters and other parameters. Then it loads data, samples hyperparameters, calls the fit function
@@ -100,7 +235,7 @@ end
 
 
 """
-	basic_experimental_loop_toy(sample_params_f, fit_f, edit_params_f, 
+	toy_experimental_loop_toy(sample_params_f, fit_f, edit_params_f, 
 		max_seed, modelname, dataset, contamination, savepath)
 
 Experimental loop for toy dataset.
